Removed the Card and Playerhand classes from the Deck class into their own standalone classes. Added more javadocs as well as a sort method for Deck.
